Move cookie storage helper to extensions
