+++ WooCommerce/Classes/Authentication/Jetpack Setup/Site Credential Login/SiteCredentialLoginViewModel.swift
@@ -144,20 +144,6 @@ private extension SiteCredentialLoginViewModel {
     }
 }
 
-extension HTTPCookieStorage {
-    func removeCookies(forHostOf url: URL) {
-        guard let siteHost = url.host?.lowercased() else { return }
-        cookies?.forEach { cookie in
-            let cookieDomain = cookie.domain
-                .trimmingCharacters(in: CharacterSet(charactersIn: "."))
-                .lowercased()
-            if siteHost == cookieDomain || siteHost.hasSuffix(".\(cookieDomain)") {
-                deleteCookie(cookie)
-            }
-        }
-    }
-}
-
 extension SiteCredentialLoginViewModel {
     enum Localization {
         static let wrongCredentials = NSLocalizedString(
